model/requirements.txt
grpcio
twilio
google-cloud-translate==2.0.1
six